Deduplicate handle_result and split large files
