+++ src/websocket/mod.rs
@@ -1,4 +1,5 @@
 pub mod connection_manager;
+pub mod correlation;
 pub mod proxy_handler;
 
 use actix_web::{web, HttpRequest, HttpResponse, Result as ActixResult};
