+++ src/websocket/proxy_handler.rs
@@ -2,11 +2,10 @@
 use actix_web::{web, Error, HttpRequest, HttpResponse};
 use actix_ws::{Message as WsMessage, MessageStream, Session};
 use futures_util::{SinkExt, StreamExt};
-use serde_json::{json, Value};
 use std::collections::HashMap;
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
-use std::time::{Instant, SystemTime, UNIX_EPOCH};
+use std::time::{SystemTime, UNIX_EPOCH};
 use tokio::sync::Mutex;
 use tokio::time::{timeout, Duration};
 use tokio_tungstenite::tungstenite::Message as TungsteniteMessage;
@@ -14,224 +13,13 @@ use tracing::{debug, error, info, warn};
 use uuid::Uuid;
 
 use super::connection_manager::WebSocketConnectionManager;
+use super::correlation::{CorrelationTracker, MessageProcessor, CORRELATION_CLEANUP_INTERVAL};
 use crate::error::AppError;
 
-const CLIENT_TIMEOUT: Duration = Duration::from_secs(300); // 5 minutes
-const MESSAGE_TIMEOUT: Duration = Duration::from_secs(30); // 30 seconds for individual messages
-const MAX_MESSAGE_SIZE: usize = 10 * 1024 * 1024; // 10MB
-const CORRELATION_TIMEOUT: Duration = Duration::from_secs(60); // 1 minute for correlation timeout
-const CORRELATION_CLEANUP_INTERVAL: Duration = Duration::from_secs(30); // 30 seconds cleanup interval
+const CLIENT_TIMEOUT: Duration = Duration::from_secs(300);
+const MESSAGE_TIMEOUT: Duration = Duration::from_secs(30);
+const MAX_MESSAGE_SIZE: usize = 10 * 1024 * 1024;
 
-/// Represents a pending request awaiting correlation with its response
-#[derive(Debug, Clone)]
-struct PendingRequest {
-    #[allow(dead_code)]
-    correlation_id: String,
-    original_message: String,
-    sent_at: Instant,
-    #[allow(dead_code)]
-    client_session_id: Uuid,
-}
-
-/// Tracks correlation state for a WebSocket session
-#[derive(Debug)]
-struct CorrelationTracker {
-    pending_requests: HashMap<String, PendingRequest>,
-    next_correlation_id: AtomicU64,
-    session_id: Uuid,
-}
-
-impl CorrelationTracker {
-    fn new(session_id: Uuid) -> Self {
-        Self {
-            pending_requests: HashMap::new(),
-            next_correlation_id: AtomicU64::new(1),
-            session_id,
-        }
-    }
-
-    fn generate_correlation_id(&self) -> String {
-        let id = self.next_correlation_id.fetch_add(1, Ordering::Relaxed);
-        let timestamp = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_millis();
-        format!("corr_{}_{}_{}", self.session_id, id, timestamp)
-    }
-
-    fn add_pending_request(&mut self, correlation_id: String, original_message: String) {
-        let request = PendingRequest {
-            correlation_id: correlation_id.clone(),
-            original_message,
-            sent_at: Instant::now(),
-            client_session_id: self.session_id,
-        };
-        debug!(
-            "Added pending request with correlation ID: {}",
-            correlation_id
-        );
-        self.pending_requests.insert(correlation_id, request);
-    }
-
-    fn remove_pending_request(&mut self, correlation_id: &str) -> Option<PendingRequest> {
-        let request = self.pending_requests.remove(correlation_id);
-        if request.is_some() {
-            debug!("Matched response with correlation ID: {}", correlation_id);
-        }
-        request
-    }
-
-    fn cleanup_expired_requests(&mut self) -> Vec<PendingRequest> {
-        let now = Instant::now();
-        let mut expired = Vec::new();
-
-        self.pending_requests.retain(|correlation_id, request| {
-            if now.duration_since(request.sent_at) > CORRELATION_TIMEOUT {
-                warn!("Correlation timeout for request: {}", correlation_id);
-                expired.push(request.clone());
-                false
-            } else {
-                true
-            }
-        });
-
-        expired
-    }
-
-    fn pending_count(&self) -> usize {
-        self.pending_requests.len()
-    }
-}
-
-/// Message processing utilities for correlation tracking
-struct MessageProcessor;
-
-impl MessageProcessor {
-    /// Attempts to inject a correlation ID into a JSON message
-    fn inject_correlation_id(
-        message: &str,
-        correlation_id: &str,
-    ) -> Result<String, serde_json::Error> {
-        // Try to parse as JSON
-        match serde_json::from_str::<Value>(message) {
-            Ok(mut json) => {
-                // Inject correlation ID into the message
-                if let Some(obj) = json.as_object_mut() {
-                    obj.insert("_correlation_id".to_string(), json!(correlation_id));
-                    debug!(
-                        "Injected correlation ID {} into JSON message",
-                        correlation_id
-                    );
-                } else {
-                    // If it's not an object, wrap it
-                    json = json!({
-                        "_correlation_id": correlation_id,
-                        "_original_message": json
-                    });
-                    debug!(
-                        "Wrapped non-object JSON with correlation ID {}",
-                        correlation_id
-                    );
-                }
-                serde_json::to_string(&json)
-            }
-            Err(_) => {
-                // Not valid JSON, try to wrap as text message
-                let wrapped = json!({
-                    "_correlation_id": correlation_id,
-                    "_original_text": message,
-                    "_wrapped": true
-                });
-                serde_json::to_string(&wrapped)
-            }
-        }
-    }
-
-    /// Attempts to extract correlation ID from a message
-    fn extract_correlation_id(message: &str) -> Option<String> {
-        match serde_json::from_str::<Value>(message) {
-            Ok(json) => {
-                if let Some(obj) = json.as_object() {
-                    // Check for correlation ID field
-                    if let Some(corr_id) = obj.get("_correlation_id") {
-                        if let Some(id_str) = corr_id.as_str() {
-                            debug!("Extracted correlation ID {} from response", id_str);
-                            return Some(id_str.to_string());
-                        }
-                    }
-
-                    // Also check common gRPC/API response patterns
-                    if let Some(corr_id) =
-                        obj.get("correlation_id").or_else(|| obj.get("request_id"))
-                    {
-                        if let Some(id_str) = corr_id.as_str() {
-                            debug!("Extracted correlation ID {} from response field", id_str);
-                            return Some(id_str.to_string());
-                        }
-                    }
-                }
-                None
-            }
-            Err(_) => None,
-        }
-    }
-
-    /// Checks if a message appears to be a request (heuristic)
-    fn is_request_message(message: &str) -> bool {
-        match serde_json::from_str::<Value>(message) {
-            Ok(json) => {
-                if let Some(obj) = json.as_object() {
-                    // Common patterns for requests
-                    obj.contains_key("method") ||
-                    obj.contains_key("command") ||
-                    obj.contains_key("action") ||
-                    obj.contains_key("request") ||
-                    // gRPC-style patterns
-                    message.contains("Request") ||
-                    // API endpoint patterns
-                    obj.contains_key("endpoint") ||
-                    obj.contains_key("path")
-                } else {
-                    false
-                }
-            }
-            Err(_) => {
-                // For non-JSON messages, use simple heuristics
-                message.contains("request") || message.contains("cmd") || message.contains("call")
-            }
-        }
-    }
-
-    /// Checks if a message appears to be a response (heuristic)
-    #[allow(dead_code)]
-    fn is_response_message(message: &str) -> bool {
-        match serde_json::from_str::<Value>(message) {
-            Ok(json) => {
-                if let Some(obj) = json.as_object() {
-                    // Common patterns for responses
-                    obj.contains_key("result") ||
-                    obj.contains_key("response") ||
-                    obj.contains_key("data") ||
-                    obj.contains_key("error") ||
-                    obj.contains_key("status") ||
-                    // gRPC-style patterns
-                    message.contains("Response") ||
-                    message.contains("Reply")
-                } else {
-                    false
-                }
-            }
-            Err(_) => {
-                // For non-JSON messages, use simple heuristics
-                message.contains("response")
-                    || message.contains("result")
-                    || message.contains("reply")
-            }
-        }
-    }
-}
-
-/// Handles WebSocket proxy connections between clients and the tapd backend
 pub struct WebSocketProxyHandler {
     connection_manager: Arc<WebSocketConnectionManager>,
     active_proxies: Arc<Mutex<HashMap<Uuid, ProxySession>>>,
@@ -943,109 +731,6 @@ mod tests {
             .await;
 
         assert_eq!(handler.active_session_count().await, 0);
-    }
-
-    #[tokio::test]
-    async fn test_correlation_tracker() {
-        let session_id = Uuid::new_v4();
-        let mut tracker = CorrelationTracker::new(session_id);
-
-        // Test correlation ID generation
-        let id1 = tracker.generate_correlation_id();
-        let id2 = tracker.generate_correlation_id();
-        assert_ne!(id1, id2);
-        assert!(id1.contains(&session_id.to_string()));
-
-        // Test adding and removing pending requests
-        tracker.add_pending_request(id1.clone(), "test message".to_string());
-        assert_eq!(tracker.pending_count(), 1);
-
-        let removed = tracker.remove_pending_request(&id1);
-        assert!(removed.is_some());
-        assert_eq!(removed.unwrap().original_message, "test message");
-        assert_eq!(tracker.pending_count(), 0);
-
-        // Test removing non-existent request
-        let removed = tracker.remove_pending_request("non-existent");
-        assert!(removed.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_message_processor_json_injection() {
-        // Test injecting correlation ID into JSON object
-        let json_message = r#"{"method": "test", "params": {"key": "value"}}"#;
-        let correlation_id = "test-corr-123";
-
-        let result = MessageProcessor::inject_correlation_id(json_message, correlation_id);
-        assert!(result.is_ok());
-
-        let modified = result.unwrap();
-        assert!(modified.contains("_correlation_id"));
-        assert!(modified.contains(correlation_id));
-
-        // Test extracting correlation ID
-        let extracted = MessageProcessor::extract_correlation_id(&modified);
-        assert_eq!(extracted, Some(correlation_id.to_string()));
-    }
-
-    #[tokio::test]
-    async fn test_message_processor_non_json() {
-        // Test with non-JSON message
-        let text_message = "This is not JSON";
-        let correlation_id = "test-corr-456";
-
-        let result = MessageProcessor::inject_correlation_id(text_message, correlation_id);
-        assert!(result.is_ok());
-
-        let modified = result.unwrap();
-        assert!(modified.contains("_correlation_id"));
-        assert!(modified.contains(correlation_id));
-        assert!(modified.contains("_original_text"));
-    }
-
-    #[tokio::test]
-    async fn test_message_type_detection() {
-        // Test request detection
-        let request_json = r#"{"method": "get_info", "params": {}}"#;
-        assert!(MessageProcessor::is_request_message(request_json));
-
-        let request_text = "send request to server";
-        assert!(MessageProcessor::is_request_message(request_text));
-
-        // Test response detection
-        let response_json = r#"{"result": {"status": "ok"}, "error": null}"#;
-        assert!(MessageProcessor::is_response_message(response_json));
-
-        let response_text = "response from server";
-        assert!(MessageProcessor::is_response_message(response_text));
-
-        // Test non-matching message
-        let other_message = r#"{"notification": "update"}"#;
-        assert!(!MessageProcessor::is_request_message(other_message));
-        assert!(!MessageProcessor::is_response_message(other_message));
-    }
-
-    #[tokio::test]
-    async fn test_correlation_timeout_cleanup() {
-        let session_id = Uuid::new_v4();
-        let mut tracker = CorrelationTracker::new(session_id);
-
-        // Add a request that should be expired
-        let correlation_id = tracker.generate_correlation_id();
-        tracker.add_pending_request(correlation_id.clone(), "test message".to_string());
-
-        // Manually set the sent_at time to be in the past
-        if let Some(request) = tracker.pending_requests.get_mut(&correlation_id) {
-            request.sent_at = Instant::now() - Duration::from_secs(120); // 2 minutes ago
-        }
-
-        assert_eq!(tracker.pending_count(), 1);
-
-        // Run cleanup
-        let expired = tracker.cleanup_expired_requests();
-        assert_eq!(expired.len(), 1);
-        assert_eq!(tracker.pending_count(), 0);
-        assert_eq!(expired[0].correlation_id, correlation_id);
     }
 
     #[tokio::test]
